Check target workspace exists or not
